Code3/requirements.txt
fastapi
uvicorn
streamlit
opencv-python
# detectron2
requests
pillow